@@ -20,51 +21,47 @@
   "dependencies": {
     "@aws-crypto/sha256-browser": "3.0.0",
     "@aws-crypto/sha256-js": "3.0.0",
-    "@aws-sdk/config-resolver": "3.272.0",
-    "@aws-sdk/fetch-http-handler": "3.272.0",
-    "@aws-sdk/hash-node": "3.272.0",
-    "@aws-sdk/invalid-dependency": "3.272.0",
-    "@aws-sdk/middleware-content-length": "3.272.0",
-    "@aws-sdk/middleware-endpoint": "3.272.0",
-    "@aws-sdk/middleware-host-header": "3.272.0",
-    "@aws-sdk/middleware-logger": "3.272.0",
-    "@aws-sdk/middleware-recursion-detection": "3.272.0",
-    "@aws-sdk/middleware-retry": "3.272.0",
-    "@aws-sdk/middleware-serde": "3.272.0",
-    "@aws-sdk/middleware-stack": "3.272.0",
-    "@aws-sdk/middleware-user-agent": "3.272.0",
-    "@aws-sdk/node-config-provider": "3.272.0",
-    "@aws-sdk/node-http-handler": "3.272.0",
-    "@aws-sdk/protocol-http": "3.272.0",
-    "@aws-sdk/smithy-client": "3.272.0",
-    "@aws-sdk/types": "3.272.0",
-    "@aws-sdk/url-parser": "3.272.0",
-    "@aws-sdk/util-base64": "3.208.0",
-    "@aws-sdk/util-body-length-browser": "3.188.0",
-    "@aws-sdk/util-body-length-node": "3.208.0",
-    "@aws-sdk/util-defaults-mode-browser": "3.272.0",
-    "@aws-sdk/util-defaults-mode-node": "3.272.0",
-    "@aws-sdk/util-endpoints": "3.272.0",
-    "@aws-sdk/util-retry": "3.272.0",
-    "@aws-sdk/util-user-agent-browser": "3.272.0",
-    "@aws-sdk/util-user-agent-node": "3.272.0",
-    "@aws-sdk/util-utf8": "3.254.0",
-    "tslib": "^2.3.1"
+    "@aws-sdk/config-resolver": "3.347.0",
+    "@aws-sdk/fetch-http-handler": "3.347.0",
+    "@aws-sdk/hash-node": "3.347.0",
+    "@aws-sdk/invalid-dependency": "3.347.0",
+    "@aws-sdk/middleware-content-length": "3.347.0",
+    "@aws-sdk/middleware-endpoint": "3.347.0",
+    "@aws-sdk/middleware-host-header": "3.347.0",
+    "@aws-sdk/middleware-logger": "3.347.0",
+    "@aws-sdk/middleware-recursion-detection": "3.347.0",
+    "@aws-sdk/middleware-retry": "3.347.0",
+    "@aws-sdk/middleware-serde": "3.347.0",
+    "@aws-sdk/middleware-stack": "3.347.0",
+    "@aws-sdk/middleware-user-agent": "3.347.0",
+    "@aws-sdk/node-config-provider": "3.347.0",
+    "@aws-sdk/node-http-handler": "3.348.0",
+    "@aws-sdk/smithy-client": "3.347.0",
+    "@aws-sdk/types": "3.347.0",
+    "@aws-sdk/url-parser": "3.347.0",
+    "@aws-sdk/util-base64": "3.310.0",
+    "@aws-sdk/util-body-length-browser": "3.310.0",
+    "@aws-sdk/util-body-length-node": "3.310.0",
+    "@aws-sdk/util-defaults-mode-browser": "3.347.0",
+    "@aws-sdk/util-defaults-mode-node": "3.347.0",
+    "@aws-sdk/util-endpoints": "3.347.0",
+    "@aws-sdk/util-retry": "3.347.0",
+    "@aws-sdk/util-user-agent-browser": "3.347.0",
+    "@aws-sdk/util-user-agent-node": "3.347.0",
+    "@aws-sdk/util-utf8": "3.310.0",
+    "@smithy/protocol-http": "^1.0.1",
+    "@smithy/types": "^1.0.0",
+    "tslib": "^2.5.0"
   },
   "devDependencies": {
-    "@aws-sdk/service-client-documentation-generator": "3.208.0",
+    "@aws-sdk/service-client-documentation-generator": "3.310.0",
     "@tsconfig/node14": "1.0.3",
     "@types/node": "^14.14.31",
     "concurrently": "7.0.0",
     "downlevel-dts": "0.10.1",
     "rimraf": "3.0.2",
-    "typedoc": "0.19.2",
-    "typescript": "~4.6.2"
-  },
-  "overrides": {
-    "typedoc": {
-      "typescript": "~4.6.2"
-    }
+    "typedoc": "0.23.23",
+    "typescript": "~4.9.5"
   },
   "engines": {
     "node": ">=14.0.0"
