@@ -1,6 +1,6 @@
 {
   "name": "@aws-sdk/config-resolver",
-  "version": "3.272.0",
+  "version": "3.347.0",
   "scripts": {
     "build": "concurrently 'yarn:build:cjs' 'yarn:build:es' 'yarn:build:types'",
     "build:cjs": "tsc -p tsconfig.cjs.json",
@@ -20,20 +20,19 @@
   },
   "license": "Apache-2.0",
   "dependencies": {
-    "@aws-sdk/signature-v4": "3.272.0",
-    "@aws-sdk/types": "3.272.0",
-    "@aws-sdk/util-config-provider": "3.208.0",
-    "@aws-sdk/util-middleware": "3.272.0",
-    "tslib": "^2.3.1"
+    "@aws-sdk/types": "3.347.0",
+    "@aws-sdk/util-config-provider": "3.310.0",
+    "@aws-sdk/util-middleware": "3.347.0",
+    "tslib": "^2.5.0"
   },
   "devDependencies": {
-    "@aws-sdk/node-config-provider": "3.272.0",
+    "@aws-sdk/node-config-provider": "3.347.0",
     "@tsconfig/recommended": "1.0.1",
     "concurrently": "7.0.0",
     "downlevel-dts": "0.10.1",
     "rimraf": "3.0.2",
-    "typedoc": "0.19.2",
-    "typescript": "~4.6.2"
+    "typedoc": "0.23.23",
+    "typescript": "~4.9.5"
   },
   "engines": {
     "node": ">=14.0.0"
@@ -46,12 +45,15 @@
     }
   },
   "files": [
-    "dist-*"
+    "dist-*/**"
   ],
   "homepage": "https://github.com/aws/aws-sdk-js-v3/tree/main/packages/config-resolver",
   "repository": {
     "type": "git",
     "url": "https://github.com/aws/aws-sdk-js-v3.git",
     "directory": "packages/config-resolver"
+  },
+  "typedoc": {
+    "entryPoint": "src/index.ts"
   }
 }
