@@ -1,6 +1,6 @@
 {
   "name": "@aws-sdk/credential-provider-env",
-  "version": "3.272.0",
+  "version": "3.347.0",
   "description": "AWS credential provider that sources credentials from known environment variables",
   "main": "./dist-cjs/index.js",
   "module": "./dist-es/index.js",
@@ -24,9 +24,9 @@
   },
   "license": "Apache-2.0",
   "dependencies": {
-    "@aws-sdk/property-provider": "3.272.0",
-    "@aws-sdk/types": "3.272.0",
-    "tslib": "^2.3.1"
+    "@aws-sdk/property-provider": "3.347.0",
+    "@aws-sdk/types": "3.347.0",
+    "tslib": "^2.5.0"
   },
   "devDependencies": {
     "@tsconfig/recommended": "1.0.1",
@@ -34,8 +34,8 @@
     "concurrently": "7.0.0",
     "downlevel-dts": "0.10.1",
     "rimraf": "3.0.2",
-    "typedoc": "0.19.2",
-    "typescript": "~4.6.2"
+    "typedoc": "0.23.23",
+    "typescript": "~4.9.5"
   },
   "types": "./dist-types/index.d.ts",
   "engines": {
@@ -49,12 +49,15 @@
     }
   },
   "files": [
-    "dist-*"
+    "dist-*/**"
   ],
   "homepage": "https://github.com/aws/aws-sdk-js-v3/tree/main/packages/credential-provider-env",
   "repository": {
     "type": "git",
     "url": "https://github.com/aws/aws-sdk-js-v3.git",
     "directory": "packages/credential-provider-env"
+  },
+  "typedoc": {
+    "entryPoint": "src/index.ts"
   }
 }
