@@ -1,27 +1,38 @@
-import React from "react";
+import React ,{useEffect}from "react";
 import { Modal, Button } from "react-bootstrap";
 import PopUpInputField from "./PopUpInputField";
 import PopUpSelection from "./PopUpSelection";
+import {MdTableRows} from 'react-icons/md';
 import { forwardRef,useImperativeHandle,useState } from "react";
-
+import axios from "axios";
+import { useParams } from "react-router-dom";
+import { Lan } from "@mui/icons-material";
 
 function ModalDialog(props,ref) {
   const [toggleOneModal, setToggleOneModal]  = React.useState(false);
   const [toggleTwoModal, setToggleTwoModal]  = React.useState(false);
   const [modalOneDataSet,setModalOneDataSet] = React.useState({});
   const [modalTwoDataSet,setModalTwoDataSet] = React.useState({});
+  const [modalOneGeneralDataSet,setModalOneGeneralDataSet] = React.useState({});
   const [editStatus,setEditStatus] = React.useState('false');
+  const [commandBasedFields,setCommandBasedFields]=useState([[],""]);
+  const [commandSet,setCommandSet]=useState([]);
+  const [dataError,setDataError]=useState("");
+  const [locatorError,setLocatorError]=useState("");
+  const [branchError,setBranchError]=useState("");
+  const [locNameErr,setLocNameErr]=useState("");
+  const [locValue,setLocValue]=useState("");
 
   let inputFieldArrayModalOne = [];
   let inputFieldArrayModalTwo = [];
   let btnValue;
   const getEditTestStep=props.editTestStep;
-  console.log('car car car');
-  console.log(getEditTestStep);
-
+  
   const testStepsData = {};
   const testStepsData2 = {};
+  const generalPurposeInputData={};
 
+  const {lname}=useParams();
   // useImperativeHandle(ref,()=> ({
   //   log(){
   //     initModalOne();
@@ -30,7 +41,7 @@ function ModalDialog(props,ref) {
   // }));
 
   const inputHandler = (name, value) => {
-    console.log(name, value);
+    console.log("ppppppppp",name, value);
     switch (name) {
       case "group":
         testStepsData[name] = value;
@@ -42,9 +53,66 @@ function ModalDialog(props,ref) {
         testStepsData[name] = value;
         break;
     }
-  
-    console.log(testStepsData);
   };
+
+  const getCommands= () => {
+    axios
+      .get("http://localhost:5000/settings/commands")
+      .then((res) => {
+        setCommandSet(res.data.settingItem); 
+        console.log("gft");
+      })
+      .catch((err) => {
+        console.log(err);
+      });
+  };
+
+  useEffect(() => {
+    getCommands();
+  }, []);
+
+
+  useEffect(() => {
+    console.log("aaaaaaaa")
+    let matchedBinaryValue;
+    for(let i=0;i<commandSet.length;i++){
+      console.log("bbbbbbb")
+      if(commandSet[i].name==modalOneDataSet["command"]){
+        matchedBinaryValue=commandSet[i].binaryValue;        ;
+      }
+    }
+    console.log("hloo",matchedBinaryValue)
+    switch(matchedBinaryValue){
+      case "000":
+        setCommandBasedFields([["swapResult","action","comment"],"000"]);
+        break;
+      case "001":
+        setCommandBasedFields([["swapResult","branchSelection","action","comment"],"001"]);  
+        break;
+      case "010":
+        setCommandBasedFields([["data","swapResult","action","comment"],"010"]);  
+        break;
+      case "011":
+        setCommandBasedFields([["data","swapResult","branchSelection","action","comment"],"011"]);  
+        break;
+      case "100":
+        setCommandBasedFields([["locator","locatorParameter","swapResult","action","comment"],"100"]);  
+        break;
+      case "101":
+        setCommandBasedFields([["locator","locatorParameter","swapResult","branchSelection","action","comment"],"101"]);  
+        break;
+      case "110":
+        setCommandBasedFields([["locator","locatorParameter","data","swapResult","action","comment"],"110"]);  
+        break;
+      case "111":
+        setCommandBasedFields([["locator","locatorParameter","data","swapResult","branchSelection","action","comment"],"111"]);  
+        break;
+
+      default:
+        setCommandBasedFields([[],""]);
+    }
+  }, [modalOneDataSet["command"]]);
+
 
   const inputHandler2 = (name, value) => {
     console.log(name, value);
@@ -72,30 +140,48 @@ function ModalDialog(props,ref) {
         break;
     }
   
-    console.log(testStepsData2);
+    console.log("ppppppppppp",testStepsData2);
   };
+
+  const inputHandler3 = (name,value) => {
+      console.log('inputHandler3: '+name,value);
+      generalPurposeInputData[name]=value;
+      console.log('inputHandler3:',props.purpose);
+  }
 
   const myLoop = () => {
     for (let i = 0; i < props.noFields[0]; i++) {
-      if(props.title[i]!=='instruction' && props.title[i]!=='command' && props.title[i]!=='swapResult' && props.title[i]!=='action'){
+      if(props.title[i]!=='instruction' && props.title[i]!=='command' && props.title[i]!=='swapResult' && props.title[i]!=='action' && props.generalPurpose===false){
         inputFieldArrayModalOne.push(
           <PopUpInputField
             id={i}
-            editStatus={editStatus}
-            editTestStep={getEditTestStep}
             title={props.title[i]}
             inputType="text"
+            generalPurpose={props.generalPurpose}
             onDataChange={inputHandler}
           ></PopUpInputField>
         );
-        console.log("Bye");
-      }else{
+      }else if(props.generalPurpose===true){//section for all general purpose data inputs such as data section,login,locator section ect...
+        console.log('jazz ',props.title);
+        console.log('jazz weke');
+        inputFieldArrayModalOne.push(
+          <PopUpInputField
+            id={i}
+            title={props.title[i]}
+            inputType="text"
+            generalPurpose={props.generalPurpose}
+            onDataChange={inputHandler3}
+            locNameErr={locNameErr}
+            locValue={locValue}
+          ></PopUpInputField>
+        ); 
+        
+      }else if(props.generalPurpose===false){
          inputFieldArrayModalOne.push(
              <PopUpSelection
              id={i}
-             editStatus={editStatus}
-             //editTestStep={props.editTestStep[props.title[i]]}
              title={props.title[i]}
+             generalPurpose={props.generalPurpose}
              onDataChange={inputHandler}
              ></PopUpSelection>
          );
@@ -103,28 +189,26 @@ function ModalDialog(props,ref) {
       
     }
     if (props.enableChainPopUps) {
-      console.log("hello koola");
-      for (let i = 0; i < props.noFields[1]; i++) {
-          if(props.title[props.noFields[0] + i]!=='instruction' && props.title[props.noFields[0] + i]!=='command' && props.title[props.noFields[0] + i]!=='swapResult' && props.title[props.noFields[0] + i]!=='action'){
-            console.log("Hi");
+      for (let i = 0; i < commandBasedFields[0].length; i++) {
+          if(commandBasedFields[0][i]!=='instruction' && commandBasedFields[0][i]!=='command' && commandBasedFields[0][i]!=='swapResult' && commandBasedFields[0][i]!=='action'){
             inputFieldArrayModalTwo.push(
               <PopUpInputField
-                id={props.noFields[0] + i}
-                editStatus={editStatus}
-                //editTestStep={props.editTestStep[props.title[props.noFields[0] + i]]}
-                title={props.title[props.noFields[0] + i]}
+                id={3+i}
+                title={commandBasedFields[0][i]}
                 inputType="text"
-                //onSaveAddFormData={addFormDataHandler}
+                generalPurpose={props.generalPurpose}
                 onDataChange2={inputHandler2}
+                dataError={dataError}
+                locatorError={locatorError}
+                branchError={branchError}
               ></PopUpInputField>
             );
           }else{
              inputFieldArrayModalTwo.push(
                <PopUpSelection
-               id={props.noFields[0] + i}
-               editStatus={editStatus}
-               //editTestStep={props.editTestStep[props.title[props.noFields[0] + i]]}
-               title={props.title[props.noFields[0] + i]}
+               id={3+i}
+               title={commandBasedFields[0][i]}
+               generalPurpose={props.generalPurpose}
                onDataChange2={inputHandler2}
                ></PopUpSelection>
             );
@@ -132,10 +216,34 @@ function ModalDialog(props,ref) {
    
       }
     }
+    // if (props.enableChainPopUps) {
+    //   for (let i = 0; i < props.noFields[1]; i++) {
+    //       if(props.title[props.noFields[0] + i]!=='instruction' && props.title[props.noFields[0] + i]!=='command' && props.title[props.noFields[0] + i]!=='swapResult' && props.title[props.noFields[0] + i]!=='action'){
+    //         inputFieldArrayModalTwo.push(
+    //           <PopUpInputField
+    //             id={props.noFields[0] + i}
+    //             title={props.title[props.noFields[0] + i]}
+    //             inputType="text"
+    //             generalPurpose={props.generalPurpose}
+    //             onDataChange2={inputHandler2}
+    //           ></PopUpInputField>
+    //         );
+    //       }else{
+    //          inputFieldArrayModalTwo.push(
+    //            <PopUpSelection
+    //            id={props.noFields[0] + i}
+    //            title={props.title[props.noFields[0] + i]}
+    //            generalPurpose={props.generalPurpose}
+    //            onDataChange2={inputHandler2}
+    //            ></PopUpSelection>
+    //         );
+    //       }
+   
+    //   }
+    // }
 
   };
   
-  console.log('eliye');
 
   myLoop();
   const initModalOne = () => {
@@ -147,14 +255,97 @@ function ModalDialog(props,ref) {
   const initModalTwo = () => {
     return setToggleTwoModal(true);
   };
-  const TerminateModalTwo = () => {
-    console.log(testStepsData2)
-    setModalTwoDataSet(Object.assign(modalOneDataSet,testStepsData2));
+
+  const closeModalTwo=()=>{
     return setToggleTwoModal(false);
+  }
+  const TerminateModalTwo = () => {
+    // setModalTwoDataSet(Object.assign(modalOneDataSet,testStepsData2));
+    console.log(modalTwoDataSet)
+    const {data,locator,branchSelection}=testStepsData2;
+    switch(commandBasedFields[1]){
+      // case "001":
+      //   if(branchSelection==undefined){
+          
+      //     return;
+      //   }
+      //   // break;
+      // case "010":
+      //   if(locator==undefined){
+      //     setLocatorError("Locator cannot be empty");
+      //     return;
+      //   }
+      //   // break;
+      case "011":
+        if(branchSelection===undefined && data!==undefined){
+          setModalTwoDataSet({...modalTwoDataSet,"data":data})
+        if(data===undefined && branchSelection!==undefined){
+          setModalTwoDataSet({...modalTwoDataSet,"branchSelection":branchSelection})
+        }
+        if(data!==undefined && branchSelection!==undefined){
+          setModalTwoDataSet(testStepsData2);
+        }
+        return;
+      }
+      // case "100":
+      //   if(data==undefined){
+      //     setDataError("Data cannot be empty");
+      //     return;
+      //   }
+      //   // break;
+      case "101":
+        if(data==undefined){
+          setDataError("Data cannot be empty");
+        }
+        if(branchSelection==undefined){
+          setBranchError("Branch Selection be empty");
+        }
+        break;
+      case "110":
+        if(data==undefined){
+          setDataError("Data cannot be empty");
+        }
+        if(locator==undefined){
+          setLocatorError("Locator cannot be empty");
+        }
+        break;
+      case "111":
+        if(data==undefined){
+          setDataError("Data cannot be empty");
+        }
+        if(locator==undefined){
+          setLocatorError("Locator cannot be empty");
+        }
+        if(branchSelection==undefined){
+          setBranchError("Branch Selection be empty");
+        }
+        break;
+    
+        }  // return setToggleTwoModal(false);
   };
 
   const NextStep = () => {
-    setModalOneDataSet(testStepsData);
+    if(props.generalPurpose===false){
+      setModalOneDataSet(testStepsData);
+    }
+    if(props.generalPurpose===true){
+      console.log('Next step',props.purpose);
+
+      if(lname[0]==='L'){
+        const {"Locator Name":name,"Locator Value":value}=generalPurposeInputData
+        if(name===undefined && value !==undefined){
+          setModalOneGeneralDataSet({...modalOneGeneralDataSet,"Locator Value":value})
+        }
+        if(value===undefined && name !==undefined){
+          setModalOneGeneralDataSet({...modalOneGeneralDataSet,"Locator Name":name})
+        }
+        if(value!==undefined && name!==undefined){
+          setModalOneGeneralDataSet(generalPurposeInputData);
+        }
+        return;
+      }
+      setModalOneGeneralDataSet(generalPurposeInputData);
+    }
     TerminateModalOne();
     if (props.enableChainPopUps) {
       setTimeout(() => {
@@ -175,7 +366,39 @@ function ModalDialog(props,ref) {
   const submitHandlerOne = (event) => {
     event.preventDefault();
     if(props.enableChainPopUps===false){
-    props.saveNewData(modalOneDataSet);
+        if(props.generalPurpose===false){
+          props.saveNewData(modalOneDataSet);
+        }
+        if(props.generalPurpose===true){
+          if(props.purpose==='fillData'){
+            setLocNameErr("");
+            setLocValue("");
+            console.log('fillData AX1');
+            if(lname[0]==="L"){
+                const {"Locator Name":name,"Locator Value":value}=modalOneGeneralDataSet
+                console.log(modalOneGeneralDataSet)
+                if(name===undefined || name===""){
+                  setLocNameErr("Locator name cannot be empty")
+                  console.log("aaaaa")
+                }
+                if(value===undefined|| value===""){
+                  setLocValue("Locator Value cannot be empty")
+                  console.log("aaaaa")
+                }
+                if(name==undefined || name==="" || value===undefined|| value===""){
+                  return;
+                }
+                console.log("eeee",name,value)
+                props.saveNewGeneralData(modalOneGeneralDataSet);// calling from heading component
+                setModalOneGeneralDataSet({})
+            }
+          }
+          if(props.purpose==='addHeading'){
+            console.log('triple H')
+            props.saveNewHeadingData(modalOneGeneralDataSet);
+          }          
+        }
+    
     TerminateModalOne();
     }
   };
@@ -183,19 +406,96 @@ function ModalDialog(props,ref) {
   const submitHandlerTwo = (event) => {
     event.preventDefault();
     if(props.enableChainPopUps===true){
-    props.saveNewData(modalTwoDataSet);
-    TerminateModalOne();
-    }
+      setDataError("");
+      setLocatorError("");
+      setBranchError("");
+      const {data,locator,branchSelection}=modalTwoDataSet;
+      console.log(modalTwoDataSet)
+      switch(commandBasedFields[1]){
+        // case "001":
+        //   if(branchSelection==undefined){
+            
+        //     return;
+        //   }
+        //   // break;
+        // case "010":
+        //   if(locator==undefined){
+        //     setLocatorError("Locator cannot be empty");
+        //     return;
+        //   }
+        //   // break;
+        case "011":
+          console.log(modalTwoDataSet)
+          if(data===undefined || data===""){
+            setDataError("Data cannot be empty")
+            console.log("aaaaa")
+          }
+          if(branchSelection===undefined|| branchSelection===""){
+            setBranchError("Branch Selection cannot be empty")
+            console.log("aaaaa")
+          }
+          if(data==undefined || data==="" || branchSelection===undefined|| branchSelection===""){
+            return;
+          }
+          const allDataSet=(Object.assign(modalOneDataSet,modalTwoDataSet));
+          console.log(allDataSet)
+          props.saveNewData(allDataSet);
+          console.log("kkkkkkk")
+          setToggleTwoModal(false)
 
+        // case "100":
+        //   if(data==undefined){
+        //     setDataError("Data cannot be empty");
+        //     return;
+        //   }
+        //   // break;
+        // case "101":
+        //   if(data==undefined){
+        //     setDataError("Data cannot be empty");
+        //   }
+        //   if(branchSelection==undefined){
+        //     setBranchError("Branch Selection be empty");
+        //   }
+        //   break;
+        // case "110":
+        //   if(data==undefined){
+        //     setDataError("Data cannot be empty");
+        //   }
+        //   if(locator==undefined){
+        //     setLocatorError("Locator cannot be empty");
+        //   }
+        //   break;
+        // case "111":
+        //   if(data==undefined){
+        //     setDataError("Data cannot be empty");
+        //   }
+        //   if(locator==undefined){
+        //     setLocatorError("Locator cannot be empty");
+        //   }
+        //   if(branchSelection==undefined){
+        //     setBranchError("Branch Selection be empty");
+        //   }
+        //   break;
+      
+        // return setToggleTwoModal(false);
+      };
+      
+        // props.saveNewData(modalTwoDataSet);
+        // console.log("kkkkkkk")
+        // setToggleTwoModal(false)
+      }
+
+    
   };
 
   return (
     <>
       <Button variant="success" onClick={initModalOne}>
-        Add
+        <MdTableRows></MdTableRows>
+        {props.buttonValue}
       </Button>
-      <form onSubmit={submitHandlerOne} id="myFormOne">
         <Modal show={toggleOneModal} tabIndex="-1">
+        <form onSubmit={submitHandlerOne} id={props.formID[0]}>
           <Modal.Header closeButton onClick={TerminateModalOne}>
             <Modal.Title>Feed Data to Test</Modal.Title>
           </Modal.Header>
@@ -204,25 +504,25 @@ function ModalDialog(props,ref) {
             <Button variant="danger" onClick={TerminateModalOne}>
               Close
             </Button>
-            <Button variant="dark" onClick={NextStep} form="myFormOne" type="submit">
+            <Button variant="dark" onClick={NextStep} form={props.formID[0]} type="submit">
               {btnValue}
             </Button>
           </Modal.Footer>
+          </form>
         </Modal>
-      </form>
 
-      <form onSubmit={submitHandlerTwo} id="myFormTwo">
       <Modal show={toggleTwoModal} tabIndex="-1">
+      <form onSubmit={submitHandlerTwo} id={props.formID[1]}>
         <Modal.Header closeButton onClick={TerminateModalTwo}>
           <Modal.Title>Feed Data to Test</Modal.Title>
         </Modal.Header>
         <Modal.Body>{inputFieldArrayModalTwo}</Modal.Body>
         <Modal.Footer>
-          <Button variant="danger" onClick={TerminateModalTwo}>
+          <Button variant="danger" onClick={closeModalTwo}>
             Close
           </Button>
           <Button
-            form="myFormTwo"
+            form={props.formID[1]}
             type="submit"
             variant="dark"
             onClick={TerminateModalTwo}
@@ -230,8 +530,8 @@ function ModalDialog(props,ref) {
             Finish
           </Button>
         </Modal.Footer>
+        </form>
       </Modal>
-      </form>
     </>
   );
 }
