@@ -11,25 +11,31 @@ const PopUpInputField = (props) => {
       setFieldName(event.target.getAttribute("name"));
       setChangeState(true);
       console.log(fieldName);
-      console.log('Miller'+fieldName);
     }
 
     useEffect(()=>{
-        console.log("My data is"+ fieldValue);
-        console.log("Field name is "+fieldName);
-         if(props.id<=2){
+        console.log("My data is", fieldValue);
+        console.log("Field name is ", fieldName);
+        console.log('Land Rover',props.id);
+         if(props.id<=2 && props.generalPurpose===false){
+           console.log('F22-Raptor');
            props.onDataChange(fieldName,fieldValue);
-         }else{
+         }else if(props.id>2 && props.generalPurpose===false){
+           console.log('B2-stealth Bomber');
            props.onDataChange2(fieldName,fieldValue);
-         }
+         }if(props.generalPurpose===true){
+           console.log('AK47');
+          props.onDataChange(fieldName,fieldValue);
+        }
     },[fieldValue]);
 
     if(changeState===false){
-      console.log('Giya giya 1');
-      if(props.id<=2){
+      if(props.id<=2 && props.generalPurpose===false){
         props.onDataChange(fieldName,fieldValue);
-      }else{
+      }else if(props.id>2 && props.generalPurpose===false){
         props.onDataChange2(fieldName,fieldValue);
+      }if(props.generalPurpose===true){
+        props.onDataChange(fieldName,fieldValue);
       }
     }
 
