@@ -2,24 +2,19 @@ import React, { useRef,useState } from "react";
 import { Button } from "react-bootstrap";
 import { propTypes } from "react-bootstrap/esm/Image";
 import EditModalDialog from "./EditPopUpWindow";
-import {MdArrowDropUp,MdArrowDropDown} from 'react-icons/md';
+import {MdModeEdit,MdDeleteForever,MdArrowDropUp,MdArrowDropDown} from 'react-icons/md';
 import './Raw.css';
 
-const Raw = ({ testStep, rawIndex, onDelete, onEdit,onArrowClick }) => {
+const Raw = ({ testStep, rawIndex, onDelete, onEdit,onArrowClick,title,generalPurpose,enableChainPopUps}) => {
+
+  const tableDataArray=[];
 
   const modalRef=useRef();
   const editButtonHandler = () => {
-    console.log('Captian Price');
     modalRef.current.log();
-    console.log(testStep);
   };
 
-  console.log("raw eliye index"+rawIndex);
-
   const onEditHandler = (editedTableData) => {
-    console.log('PVT.Miller');
-    console.log('Index in raw hihi'+rawIndex);
-    //console.log('Index in raw props'+myindex);
     onEdit(editedTableData,rawIndex);
   }
 
@@ -34,58 +29,47 @@ const Raw = ({ testStep, rawIndex, onDelete, onEdit,onArrowClick }) => {
     }
   }
 
+  //const arrayConvertor=Object.values(testStep);//convert testStep object to an array
+
+  for(let i=0;i<title.length;i++){
+    console.log('your test step:',testStep);
+    let key=title[i];
+    if(testStep[key]===undefined || testStep[key]===" "){
+      tableDataArray.push(
+        <td>{"null"}</td>
+      )
+    }else{
+      tableDataArray.push(
+        <td>{testStep[key]}</td>
+      )
+    }
+    
+  }
+
   return (
     <>
       <EditModalDialog
-        enableChainPopUps={true}
+        enableChainPopUps={enableChainPopUps}
         ref={modalRef}
-        title={[
-          "group",
-          "instruction",
-          "command",
-          "locator",
-          "locatorParameter",
-          "data",
-          "swapResult",
-          "branchSelection",
-          "action",
-          "comment",
-        ]}
+        title={title}
         noFields={[3, 7]}
         rawNumber={null}
         raw={testStep} 
         index={rawIndex}
         onEdit={onEditHandler}
+        generalPurpose={generalPurpose}
       ></EditModalDialog>
       <tr>
-        <td>{testStep.group}</td>
-        <td>{testStep.instruction}</td>
-        <td>{testStep.command}</td>
-        <td>{testStep.locator}</td>
-        <td>{testStep.locatorParameter}</td>
-        <td>{testStep.data}</td>
-        <td>{testStep.swapResult}</td>
-        <td>{testStep.branchSelection}</td>
-        <td>{testStep.action}</td>
-        <td>{testStep.comment}</td>
-        <td><MdArrowDropUp onClick={()=>moveUpDownHandler(0)}/><MdArrowDropDown onClick={()=>moveUpDownHandler(1)}/></td>
-        <td>
+        <td className="table-data">
           {" "}
-          <Button variant="success" onClick={()=>editButtonHandler()}>
-            Edit
-          </Button>
+          <MdModeEdit color="04D9FF" size="20px"  onClick={editButtonHandler}></MdModeEdit>
         </td>
-        <td>
+        <td className="table-data">
           {" "}
-          <Button
-            variant="danger"
-            onClick={() => {
-              onDelete(rawIndex);
-            }}
-          >
-            Delete
-          </Button>
+          <MdDeleteForever color="#FF3131" size="20px" onClick={()=>{onDelete(rawIndex)}}></MdDeleteForever>
         </td>
+        <td className="table-data"><MdArrowDropUp size="25px" color="#39ff14" onClick={()=>moveUpDownHandler(0)}/><MdArrowDropDown size="25px" color="#39ff14" onClick={()=>moveUpDownHandler(1)}/></td>
+        {tableDataArray}
       </tr>
     </>
   );
