@@ -1,62 +1,59 @@
-import React from "react";
-import { useState } from "react";
-import { Modal, Button } from "react-bootstrap";
-import { forwardRef, useImperativeHandle} from "react";
+// import React from "react";
+// import { useState } from "react";
+// import { Modal, Button } from "react-bootstrap";
+// import { forwardRef, useImperativeHandle} from "react";
 
-const SettingEditPopup = ({ type, itemID ,value,callingFrom, addNewOrEdit}, ref) => {
-  let url = "http://localhost:5000/settings/" + type + "/" + itemID;
+// const SettingEditPopup = ({ type, itemID ,value, edit}, ref) => {
+//   // let url = "http://localhost:5000/settings/" + type + "/" + itemID;
 
-  const [enablePopup, setEnablePopup] = useState(false);
+//   const [enablePopup, setEnablePopup] = useState(false);
 
-  console.log(value);
+//   console.log(value);
 
-  const [fieldValue,setFieldValue]=useState(value);
-  console.log(fieldValue)
-  const inputHandler=(event) => {
-    setFieldValue(event.target.value);
-  }
+//   const [fieldValue,setFieldValue]=useState(value);
+//   console.log(fieldValue)
+//   const inputHandler=(event) => {
+//     setFieldValue(event.target.value);
+//   }
 
-  const closeModal = () => {
-    setEnablePopup(false);
-    setFieldValue("")
-  };
+//   const closeModal = () => {
+//     setEnablePopup(false);
+//     setFieldValue(value)
+//   };
 
-  useImperativeHandle(ref, () => ({
-    open() {
-      setEnablePopup(true);
-    },
-  }));
+//   useImperativeHandle(ref, () => ({
+//     open() {
+//       setEnablePopup(true);
+//     },
+//   }));
 
-  const onSubmitHandler = (event) => {
-    event.preventDefault();
-    if(callingFrom==="add"){
-      console.log('hiiiiiiiiii');
-      addNewOrEdit(fieldValue);
-    }
-    setEnablePopup(false);
-    setFieldValue("");
-  };
+//   const onSubmitHandler = (event) => {
+//     event.preventDefault();
+//     edit(fieldValue);
+//     setEnablePopup(false);
+//     setFieldValue("");
+//   };
 
-  return (
-    <form onSubmit={onSubmitHandler} id="settingsEditPopup" action="" method="POST">
-      <Modal show={enablePopup}>
-        <Modal.Header closeButton onClick={closeModal}>
-          <Modal.Title>{callingFrom} {type}</Modal.Title>
-        </Modal.Header>
-        <Modal.Body>
-          <input type="text" value={fieldValue} onChange={inputHandler}></input>
-        </Modal.Body>
-        <Modal.Footer>
-          <Button variant="danger" onClick={closeModal}>
-            Close
-          </Button>
-          <Button variant="dark" form="settingsEditPopup" type="submit">
-            {callingFrom==="add"?<span>Add</span>:<span>Edit</span>}
-          </Button>
-        </Modal.Footer>
-      </Modal>
-    </form>
-  );
-};
+//   return (
+//     <form onSubmit={onSubmitHandler} id="settingsEPopup" action="" method="POST">
+//       <Modal show={enablePopup}>
+//         <Modal.Header closeButton onClick={closeModal}>
+//           <Modal.Title>Edit {type}</Modal.Title>
+//         </Modal.Header>
+//         <Modal.Body>
+//           <input type="text" value={fieldValue} onChange={inputHandler}></input>
+//         </Modal.Body>
+//         <Modal.Footer>
+//           <Button variant="danger" onClick={closeModal}>
+//             Close
+//           </Button>
+//           <Button variant="dark" form="settingsEPopup" type="submit">
+//             <span>Edit</span>
+//           </Button>
+//         </Modal.Footer>
+//       </Modal>
+//     </form>
+//   );
+// };
 
-export default forwardRef(SettingEditPopup);
+// export default forwardRef(SettingEditPopup);
